refactor: move logo & every nav bar action to single component
